Calculate an isodistance around a French location

Queries the Geoservices (IGN) API using the /isochrone endpoint to calculate an isodistance around a French (mainland) location. This service uses the IGN BD TOPO® Valhalla to get road data.

An isodistance is an area linking the points accessible by a given means of transport (argument profile) for a given distance (argument cost_value).

The API documentation is available here.

Usage

gs_get_isodistance(
  longitude,
  latitude,
  cost_value,
  profile = "car",
  direction = "departure",
  distance_unit = "kilometer"
)

Arguments

longitude

a numeric of length 1. The longitude of the location defined in the WGS84 system. The location must belongs to Mainland France.

latitude

a numeric of length 1. The latitude of the location defined in the WGS84 system. The location must belongs to Mainland France.

cost_value

a numeric of length 1. The time used to calculate the isodistance. Unit is specified by the argument distance_unit. Must be strictly positive.

profile

a character of length 1. The means of transport used to calculate the isodistance. Possible values: car and pedestrian.

direction

a character of length 1. The course direction. If direction = "departure", the isodistance will be calculated from the starting point (defined by longitude and latitude arguments) and will provide the potential arrival points. If direction = "arrival", the isodistance will provide the potential departure points to the arrival point (defined by longitude and latitude arguments).

distance_unit

a character of length 1. The distance unit of the argument cost_value. Possible values: hour, minute and second.

Value

The isodistance as a sf POLYGON object.

Examples

if (FALSE) { # \dontrun{
# Set Florac-Trois-Rivieres, Lozere coordinates ----

florac <- data.frame(longitude  =  3.59608777,
                     latitude   = 44.32690171)

# Isodistance of 10 km around Florac-Trois-Rivieres, Lozere (by car) ----

iso <- gs_get_isodistance(longitude     = florac$"longitude", 
                          latitude      = florac$"latitude", 
                          cost_value    = 10,
                          distance_unit = "kilometer",
                          profile       = "car")

# Convert point to sf object ----

florac <- sf::st_as_sf(florac, coords = 1:2, crs = sf::st_crs(4326))

# Make map ----

library(ggplot2)

ggplot() +
  geom_sf(data = iso,  fill = "red", col = "white") +
  geom_sf(data = florac, col = "white", shape = 19) +
  theme_bw()
} # }